+++ src/util/keys.rs
@@ -5,21 +5,25 @@ use std::str::FromStr;
 /* Operational keys and combinations */
 #[derive(Debug)]
 pub struct ActionKeys {
-	pub main_key: Keycode,
-	pub opt_keys: Vec<Keycode>,
+	key_groups: Vec<Vec<Keycode>>,
 }
 
 /* Display implementation for user-facing output */
 impl fmt::Display for ActionKeys {
 	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		let mut keys = format!("{:?}-", self.main_key);
-		for (i, opt_key) in self.opt_keys.iter().enumerate() {
-			keys += &format!("{:?}", opt_key);
-			if i != self.opt_keys.len() - 1 {
-				keys += "/"
-			}
-		}
-		write!(f, "{}", keys)
+		write!(
+			f,
+			"{}",
+			self.key_groups
+				.iter()
+				.map(|keys| keys
+					.iter()
+					.map(|key| format!("{:?}", key))
+					.collect::<Vec<String>>()
+					.join("-"))
+				.collect::<Vec<String>>()
+				.join(",")
+		)
 	}
 }
 
@@ -27,8 +31,10 @@ impl fmt::Display for ActionKeys {
 impl Default for ActionKeys {
 	fn default() -> Self {
 		Self {
-			main_key: Keycode::LAlt,
-			opt_keys: vec![Keycode::S, Keycode::Enter],
+			key_groups: vec![
+				vec![Keycode::LAlt, Keycode::S],
+				vec![Keycode::LAlt, Keycode::Enter],
+			],
 		}
 	}
 }
@@ -37,12 +43,23 @@ impl ActionKeys {
 	/**
 	 * Create a new ActionKeys object.
 	 *
-	 * @param  main_key
-	 * @param  opt_keys
+	 * @param  key_groups
 	 * @return ActionKeys
 	 */
-	pub fn new(main_key: Keycode, opt_keys: Vec<Keycode>) -> Self {
-		Self { main_key, opt_keys }
+	pub fn new(key_groups: Vec<Vec<Keycode>>) -> Self {
+		Self { key_groups }
+	}
+
+	/**
+	 * Return the primary keycodes.
+	 *
+	 * @return Vector of KeyCode
+	 */
+	pub fn get_primary(&self) -> Vec<&Keycode> {
+		self.key_groups
+			.iter()
+			.filter_map(|keys| keys.get(0))
+			.collect()
 	}
 
 	/**
@@ -52,18 +69,21 @@ impl ActionKeys {
 	 * @return ActionKeys
 	 */
 	pub fn parse(keys: &str) -> Self {
-		let keys = keys.split('-').collect::<Vec<&str>>();
-		Self::new(
-			Keycode::from_str(keys.get(0).unwrap_or(&"LAlt")).expect("Invalid key"),
-			keys.get(1)
-				.unwrap_or(&"S/Enter")
-				.split('/')
-				.map(|k| {
-					Keycode::from_str(k)
-						.unwrap_or_else(|_| panic!("Invalid key ({})", k))
-				})
-				.collect(),
-		)
+		let keys = keys
+			.split(',')
+			.filter_map(|keys| {
+				let group = keys
+					.split('-')
+					.filter_map(|v| Keycode::from_str(v).ok())
+					.collect::<Vec<Keycode>>();
+				(!group.is_empty()).then(|| group)
+			})
+			.collect::<Vec<Vec<Keycode>>>();
+		if keys.is_empty() {
+			Self::default()
+		} else {
+			Self::new(keys)
+		}
 	}
 
 	/**
@@ -73,18 +93,15 @@ impl ActionKeys {
 	 * @return bool
 	 */
 	pub fn check(&self, keys: Vec<Keycode>) -> bool {
-		if !keys.contains(&self.main_key) {
-			false
-		} else {
-			let mut pressed = false;
-			for key in &self.opt_keys {
-				if keys.contains(&key) {
-					pressed = true;
-					break;
-				}
+		for target_keys in &self.key_groups {
+			if target_keys.len() == keys.len()
+				&& target_keys.len()
+					== target_keys.iter().filter(|k| keys.contains(k)).count()
+			{
+				return true;
 			}
-			pressed
 		}
+		false
 	}
 }
 
@@ -94,16 +111,42 @@ mod tests {
 	use pretty_assertions::assert_eq;
 	#[test]
 	fn test_action_keys() {
-		let keys = "LControl-Q/W";
-		let action_keys = ActionKeys::parse(keys);
-		assert_eq!(keys, action_keys.to_string());
-		assert_eq!(Keycode::LControl, action_keys.main_key);
-		assert_eq!(vec![Keycode::Q, Keycode::W], action_keys.opt_keys);
-		assert!(!action_keys.check(vec![Keycode::RAlt, Keycode::X]));
-		assert!(!action_keys.check(vec![Keycode::LControl, Keycode::X]));
-		assert!(!action_keys.check(vec![Keycode::LControl]));
-		assert!(!action_keys.check(vec![Keycode::W]));
-		assert!(action_keys.check(vec![Keycode::LControl, Keycode::Q]));
-		assert!(action_keys.check(vec![Keycode::LControl, Keycode::W]));
+		let keys_str = "LControl-Q,LControl-W";
+		let keys = ActionKeys::parse(keys_str);
+		assert_eq!(keys_str, keys.to_string());
+		assert_eq!(
+			vec![
+				vec![Keycode::LControl, Keycode::Q],
+				vec![Keycode::LControl, Keycode::W]
+			],
+			keys.key_groups
+		);
+		assert!(!keys.check(vec![Keycode::RAlt, Keycode::X]));
+		assert!(!keys.check(vec![Keycode::LControl, Keycode::X]));
+		assert!(!keys.check(vec![Keycode::LControl]));
+		assert!(!keys.check(vec![Keycode::W]));
+		assert!(keys.check(vec![Keycode::LControl, Keycode::Q]));
+		assert!(keys.check(vec![Keycode::LControl, Keycode::W]));
+		assert!(!ActionKeys::parse("S").check(vec![Keycode::S, Keycode::Slash]));
+		assert!(ActionKeys::parse("X,Y").check(vec![Keycode::X]));
+		assert!(ActionKeys::parse("LControl-J,A,B,C")
+			.check(vec![Keycode::LControl, Keycode::J]));
+		assert!(ActionKeys::parse("LControl-A,X,Y").check(vec![Keycode::Y]));
+		assert_eq!(
+			Vec::<Vec<Keycode>>::new(),
+			ActionKeys::parse("LCxntrxl-WW").key_groups
+		);
+		assert_eq!(
+			vec![vec![Keycode::X]],
+			ActionKeys::parse("test,X,...").key_groups
+		);
+		assert_eq!(
+			vec![vec![Keycode::LControl], vec![Keycode::X]],
+			ActionKeys::parse("LControl-test,X").key_groups
+		);
+		assert_eq!(
+			vec![&Keycode::A, &Keycode::C],
+			ActionKeys::parse("A-B,C-D,...").get_primary()
+		);
 	}
 }